# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API-PERSON ",
        "description": "Documentação da API-PERSON com Swagger Service.",
        "contact": {
            "name": "Luiz Santos Full Stack Developer ",
            "email": "luizcsdev@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor Local"
        }
    ],
    "paths": {
        "/enderecos": {
            "get": {
                "tags": [
                    "Endereço"
                ],
                "summary": "Lista todos os endereços",
                "description": "Retorna uma lista dos endereços armazenados no banco de dados.",
                "operationId": "736b5681b3a3623d5350cfb852357250",
                "responses": {
                    "200": {
                        "description": "Lista de endereços retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Endereco"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nenhum endereços encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Não há endereços cadastrados!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno ao buscar os endereços",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao buscar os endereços."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Detalhes do erro"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Endereço"
                ],
                "summary": "Cria um novo Endereço",
                "description": "Registra um novo Endereço no banco de dados.",
                "operationId": "649a502077b75ae60503d438b32c3adb",
                "requestBody": {
                    "description": "Dados necessários para criar um novo Endereço",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cid_id",
                                    "end_tipo_logradouro",
                                    "end_logradouro",
                                    "end_numero",
                                    "end_bairro"
                                ],
                                "properties": {
                                    "cid_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "end_tipo_logradouro": {
                                        "type": "string",
                                        "example": "Avenida"
                                    },
                                    "end_logradouro": {
                                        "type": "string",
                                        "example": "Gustavo da Silveira"
                                    },
                                    "end_numero": {
                                        "type": "integer",
                                        "example": "1000"
                                    },
                                    "end_bairro": {
                                        "type": "string",
                                        "example": "Horto Florestal"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Endereço criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Endereco criado com sucesso."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Endereco"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao criar o Endereço",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro ao criar o Endereço."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Endereco": {
                "title": "Endereço",
                "description": "Modelo de Endereço",
                "required": [
                    "cid_id",
                    "end_tipo_logradouro",
                    "end_logradouro",
                    "end_numero",
                    "end_bairro"
                ],
                "properties": {
                    "end_id": {
                        "description": "ID do Endereço",
                        "type": "integer",
                        "example": 1
                    },
                    "cid_id": {
                        "description": "ID da Cidade",
                        "type": "integer",
                        "example": "1"
                    },
                    "end_tipo_logradouro": {
                        "description": "Descricão do tipo de logradouro",
                        "type": "string",
                        "example": "Avenida"
                    },
                    "end_logradouro": {
                        "description": "Descricão do logradouro",
                        "type": "string",
                        "example": "Gustavo da Silveira"
                    },
                    "end_numero": {
                        "description": "Número do logradouro",
                        "type": "integer",
                        "example": "1000"
                    },
                    "end_bairro": {
                        "description": "Bairro do logradouro",
                        "type": "string",
                        "example": "Horto Florestal"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}